requirements.txt
Babel==2.7.0
backcall==0.1.0
boto3==1.9.217
botocore==1.12.217
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
decorator==4.4.0
dj-database-url==0.5.0
Django==2.0.13
django-bootstrap3==11.1.0
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==2.2
django-oscar==2.0.1
django-phonenumber-field==2.0.1
django-redis==4.10.0
django-s3-storage==0.12.5
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
docutils==0.15.2
easy-thumbnails==2.6
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.0
flake8==3.7.8
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
gunicorn==19.9.0
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
mccabe==0.6.1
mock==2.0.0
parso==0.5.1
pbr==5.4.2
pexpect==4.7.0
phonenumbers==8.10.16
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
purl==1.5
pyasn1==0.4.6
pycodestyle==2.5.0
pycountry==19.7.15
pyflakes==2.1.1
Pygments==2.4.2
pyprof2calltree==1.4.4
pysolr==3.7.0
python-dateutil==2.8.0
python-stdnum==1.11
pytz==2019.2
PyYAML==5.1.2
redis==3.3.6
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
sorl-thumbnail==12.4.1
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.25.3
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.15.5
whitenoise==4.1.3
Whoosh==2.7.4
